// src/app/common/interfaces/VehicleCostTracking.interface.ts
export interface VehicleCostTracking {
    VCHRNO: string;
    AD_VOUCHERNO: string;
    ACID: string;
    ACNAME: string;
    VEHICLE_NO: string;
    AMOUNT: number;
    VT_NO:string;
    CAPITALPURCHASE_VCHRNO:string;
    REFVOUCHER_TOTAL:number;
    SL_ACID:string;
    SL_ACNAME:string;
    HASSUBLEDGER:number;
}